CyberSparrow - Browser/secure_browser/requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
requests>=2.28.0
python-whois>=0.8.0
dnspython>=2.3.0
python-nmap>=0.7.1
beautifulsoup4>=4.11.0
cryptography>=39.0.0
pyyaml>=6.0
scapy>=2.5.0
aiohttp>=3.8.3
pyopenssl>=23.0.0
python-dotenv>=0.21.0
sqlalchemy>=2.0.0
pytest>=7.3.1
black>=22.3.0
pylint>=2.15.0
psutil>=5.9.0 